requirements.txt
streamlit
pandas
requests
numpy
pillow
